Learning (#169)
This commit adds learning features into the framework. It allows multi-agent learning using MATD3 algorithm. For now the learning is only available for power plants.

It also brings multiple changes and additions to the grafana dashboard. Now all parameters for learning and evaluation are tracked. Also both accepted and rejected orders are logged.

---------

Co-authored-by: Nick Harder <nick.harder94@gmail.com>
Co-authored-by: kim-mskw <kim.miskiw@kit.edu>
Co-authored-by: Florian Maurer <maurer@fh-aachen.de>
Co-authored-by: Philipp Fritz <fritzphilipp@outlook.de>