Improve speed (#357)
As suggested by @maurerle I tried to find a good solution to imrove
speed and also calculate the generation costs, but I didn't manage to
find a good solution. Since the speed-up is around 2-3x times, I am
willing to sacrifice this function until we fix it in a better way.