Tests and classes of Dst components, steel plant, flexibility script (#344)
This PR adds the Steel Plant Demand Side Management Unit. This also introduces the whole framework of DSM units and demand side components, which can be expended in the future. A detailed documentation as well as a tutorial on the use of the DSM units is to be expected soon.


Co-authored-by: Nick Harder <nick.harder94@gmail.com>
Co-authored-by: Florian Maurer <maurer@fh-aachen.de>