Improve learning logging into the database (#554)
We still have the issue, that the grafana dashboard is not usable when
running large simulations, and this is actually very useful and
important when trying to get large trainings to work, especially with
different agents. I have identified, that the problem is due to the
query of all the simulation ids and processing and so on, it just gets
stuck when I have multiple big runs in the database. Hence I have
decided to completely go away from the idea of changing simulation_id
during training and logging only using episode and mode, which improved
the performance and makes everything easier. I was also not happy that
we didn't have direct control over the episode logged and used the
simulation_id for that. This has been addressed as well

## Changes Proposed
- Changes the way rl_params are stored in the database and gets rid of
changing simulation id
- change perform_evaluation to proper name evaluation_mode
- remove changing sim_id during learning
- store rl_parals under unified simulation_id and differentiate using
episode
- use simulation_id as variable through the whole code for better
readability
- update dashboard

---------

Co-authored-by: Florian Maurer <maurer@fh-aachen.de>