correctly suppress highs output in linopy (#622)
# Pull Request

The `suppress_output` method redirects the output buffers.
This is generally not a good idea, instead, the underlying source of
unwanted output should be tackled.

This PR fixes the output from highs in the redispatch market.

## Related Issue
This fixes #619 as this produced errors in Colab.

## Description
There is still a line `Running HiGHS 1.11.0 (git hash: 364c83a):
Copyright (c) 2025 HiGHS under MIT licence terms` which is printed for
every solving.
This is going to be fixed in linopy with
https://github.com/PyPSA/linopy/pull/479

## Changes Proposed
- correctly pass the silence args to the highs solver
- delete now unused suppress_output function
- improve package installation in colab by calling pip less often

## Testing
[Describe the testing you've done, including any specific test cases or
scenarios]

## Checklist
Please check all applicable items:

- [ ] Code changes are sufficiently documented (docstrings, inline
comments, `doc` folder updates)
- [ ] New unit tests added for new features or bug fixes
- [ ] Existing tests pass with the changes
- [ ] Reinforcement learning examples are operational (for DRL-related
changes)
- [ ] Code tested with both local and Docker databases
- [ ] Code follows project style guidelines and best practices
- [ ] Changes are backwards compatible, or deprecation notices added
- [ ] New dependencies added to `pyproject.toml`
- [ ] A note for the release notes `doc/release_notes.rst` of the
upcoming release is included
- [x] Consent to release this PR's code under the GNU Affero General
Public License v3.0

## Additional Notes (if applicable)
[Any additional information, concerns, or areas you want reviewers to
focus on]

## Screenshots (if applicable)
[Add screenshots to demonstrate visual changes]