docker: fix build by pinning pip-tools (#626)
# Pull Request

There is an issue with self-referential packages in a pyproject.toml
with the latest pip-tools release.
The self-referential [all] package, which is not written to the
requirements.txt as it is not selected still throws a pip-compile error.

More information for this can be found in the ticket:
https://github.com/jazzband/pip-tools/issues/2215

The pin should be removed when
https://github.com/jazzband/pip-tools/pull/2221 gets merged.